Добавляет индексы, изменяет расположение файлов
