+++ src/api/auth/index.ts
@@ -1,0 +1,2 @@
+export { AuthApi, authApi } from "./AuthApi";
+export { SignInData, SignUpData } from "./types";
