+++ src/api/auth/types.ts
@@ -1,0 +1,24 @@
+export type SignInData = {
+  login: string;
+  password: string;
+};
+
+export type User = {
+  id: number;
+  first_name: string;
+  second_name: string;
+  display_name: string;
+  login: string;
+  email: string;
+  phone: number;
+  avatar: string;
+};
+
+export type SignUpData = {
+  first_name: string;
+  second_name: string;
+  login: string;
+  email: string;
+  password: string;
+  phone: string;
+};
