+++ src/api/index.ts
@@ -1,1 +1,1 @@
-export { AuthApi, authApi, SignInData, SignUpData } from "./AuthApi";
+export * from "./auth";
